add styleable property (#9)

+++ indicatorview/src/main/java/com/github/sasakitomohiro/indicatorview/ViewExt.kt
@@ -4,6 +4,6 @@ import android.content.res.Resources
 
 private val metrics = Resources.getSystem().displayMetrics
 
-internal fun Int.dpToPx() = (this * (metrics.densityDpi / 160f)).toInt()
+internal fun Int.dpToPx() = this * (metrics.densityDpi / 160f)
 
 internal fun Int.pxToDp() = (this / (metrics.densityDpi / 160f)).toInt()
